+++ src/rating.c
@@ -139,8 +139,8 @@ void calculate_elo_ratings(rating* p_winner, rating* p_loser)
     double expected_winner_score = 1 / ( 1 + pow(10, (p_loser->score - p_winner->score) / 400));
     double expected_loser_score = 1 / ( 1 + pow(10, (p_winner->score - p_loser->score) / 400));
 
-    int winner_change = K_FACTOR * (1 - expected_winner_score);
-    int loser_change = K_FACTOR * expected_loser_score;
+    int winner_change =  (int)(K_FACTOR * (1 - expected_winner_score));
+    int loser_change = (int)(K_FACTOR * expected_loser_score);
 
     p_winner->score += winner_change;
     p_loser->score -= loser_change;
